Add nCore
I didnt fint any info on how to determine the join diff.

On nCore:

You have to have a lot of point and also have to be a paid member to inv someone.
You are also responsibe for anyone who you invite.
So I figured Hard would be an appropriate diff.